+++ config/webpack.styles.config.js
@@ -32,6 +32,15 @@ module.exports = [
         options: {
           importLoaders: 1
         }
+      },
+      {
+        loader: 'postcss-loader',
+        options: {
+          postcssOptions: {
+            // 添加 autoprefixer 插件
+            plugins: [require('autoprefixer')]
+          }
+        }
       }
     ]
   }